requirements.txt
asgiref==3.8.1
boto3==1.37.4
botocore==1.37.4
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
Django==4.2.19
djangorestframework==3.15.2
google-api-core==2.24.1
google-api-python-client==2.162.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.68.0
httplib2==0.22.0
idna==3.10
jmespath==1.0.1
oauthlib==3.2.2
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
s3transfer==0.11.3
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==1.26.20
